+++ shapes_converter/src/shex_to_uml/value_constraint.rs
@@ -4,6 +4,7 @@ use super::Name;
 pub enum ValueConstraint {
     #[default]
     Any,
+    ValueSet(Vec<Name>),
     Datatype(Name),
     Ref(Name),
     None,